installer: Detect previous installations and more error handling.

+++ public/install.sh
@@ -46,16 +46,24 @@ stateexe=$os-$arch
 
 echo "Preparing for installation..."
 
-if [ ! -f $statepkg ]; then
+if [ ! -f $statepkg -a ! -f $stateexe ]; then
   if [ ! -z "`which wget`" ]; then
     # Determine the latest version to fetch.
     version=`wget -q -O - $STATEURL$statejson | grep -m 1 '"Version":' | awk '{print $2}' | tr -d '",'`
+    if [ -z "$version" ]; then
+      echo "Unable to retrieve the latest version number"
+      exit 1
+    fi
     echo "Fetching the latest version: $version"
     # Fetch it.
     wget -q ${STATEURL}${version}/${statepkg} || exit 1
   elif [ ! -z "`which curl`" ]; then
     # Determine the latest version to fetch.
     version=`curl -s $STATEURL$statejson | grep -m 1 '"Version":' | awk '{print $2}' | tr -d '",'`
+    if [ -z "$version" ]; then
+      echo "Unable to retrieve the latest version number"
+      exit 1
+    fi
     echo "Fetching the latest version: $version"
     # Fetch it.
     curl -s -o $statepkg ${STATEURL}${version}/${statepkg} || exit 1
@@ -66,8 +74,11 @@ if [ ! -f $statepkg ]; then
 fi
 
 # Extract the State binary.
-echo "Extracting $statepkg..."
-gunzip $statepkg || exit 1
+if [ -f $statepkg ]; then
+  echo "Extracting $statepkg..."
+  gunzip $statepkg || exit 1
+  chmod +x $stateexe
+fi
 
 # Verify checksum.
 echo "Verifying checksum..."
@@ -80,13 +91,20 @@ if [ "`sha256sum -b $stateexe | cut -d ' ' -f1`" != "$shasum" ]; then
   exit 1
 fi
 
-# Prompt the user for a directory to install to, defaulting to /usr/local/bin if
-# the user has write permissions, else a local bin.
-if [ -w "/usr/local/bin" ]; then
-  installdir="/usr/local/bin"
+# Check for existing installation. Otherwise, make the installation default to
+# /usr/local/bin if the user has write permission, or to a local bin.
+installdir="`dirname \`which $STATEEXE\``"
+if [ ! -z "$installdir" ]; then
+  echo "Previous installation detected at $installdir"
 else
-  installdir="$HOME/.local/bin"
+  if [ -w "/usr/local/bin" ]; then
+    installdir="/usr/local/bin"
+  else
+    installdir="$HOME/.local/bin"
+  fi
 fi
+
+# Prompt the user for a directory to install to.
 while "true"; do
   echo -n "Please enter the installation directory [$installdir]: "
   read input
@@ -109,6 +127,9 @@ while "true"; do
     echo "NOTE: $installdir will be created"
   elif [ -e "$installdir/$STATEEXE" ]; then
     echo "WARNING: overwriting previous installation"
+  fi
+  if [ "`dirname \`which $STATEEXE\``" != "$installdir" ]; then
+    echo "WARNING: installing elsewhere from previous installation"
   fi
   echo -n "Continue? [y/N/q] "
   read response
@@ -136,24 +157,24 @@ done
 
 # Check if the installation is in $PATH, if not, update user's profile if
 # permitted to.
-if [ "$installdir" = "`echo \"$PATH\" | grep -Fo \"$installdir\"`" ]; then
+if [ "`dirname \`which $STATEEXE\``" = "$installdir" ]; then
   echo "Installation complete."
-  echo "You may now start using the 'state' program."
+  echo "You may now start using the '$STATEEXE' program."
   exit 0
 fi
 profile="`echo $HOME`/.profile"
 if [ ! -w "$profile" ]; then
   echo "Installation complete."
-  echo "Please manually add $installdir to your \$PATH in order to start"
-  echo "using the 'state' program."
+  echo -n "Please manually add $installdir to your \$PATH in order to start "
+  echo "using the '$STATEEXE' program."
   exit 1
 fi
 echo -n "Allow \$PATH to be appended to in your $profile? [y/N]"
 read response
 if [ "$response" != "Y" -a "$response" != "y" ]; then
   echo "Installation complete."
-  echo "Please manually add $installdir to your \$PATH in order to start"
-  echo "using the 'state' program."
+  echo -n "Please manually add $installdir to your \$PATH in order to start "
+  echo "using the '$STATEEXE' program."
   exit 1
 fi
 echo "Updating environment..."
@@ -167,5 +188,5 @@ else
 fi
 
 echo "Installation complete."
-echo "Please either run 'source ~/.profile' or start a new login shell in "
-echo "order to start using the 'state' program."
+echo -n "Please either run 'source ~/.profile' or start a new login shell in "
+echo "order to start using the '$STATEEXE' program."
